analytics/flink-job/python/requirements.txt
apache-flink==1.18.1
google-cloud-pubsub==2.18.4
google-cloud-firestore==2.11.1
pyflink==1.18.1


